move: member domain 하위 디렉토리로 이동
