+++ src/main/kotlin/edu/example/kotlindevelop/domain/member/exception/MemberTaskException.kt
@@ -1,3 +1,3 @@
-package edu.example.kotlindevelop.member.exception
+package edu.example.kotlindevelop.domain.member.exception
 
 class MemberTaskException(message: String?, val statusCode: Int) : RuntimeException(message)
